@@ -1,8 +1,8 @@
 {
   "name": "@tanstack/svelte-form",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "description": "Powerful, type-safe forms for Svelte.",
-  "author": "James Garbutt (https://github.com/43081j)",
+  "author": "tannerlinsley",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -43,13 +43,14 @@
     "src"
   ],
   "dependencies": {
-    "@tanstack/form-core": "workspace:*"
+    "@tanstack/form-core": "workspace:*",
+    "@tanstack/svelte-store": "^0.7.0"
   },
   "devDependencies": {
-    "@sveltejs/package": "^2.3.7",
-    "@sveltejs/vite-plugin-svelte": "^4.0.4",
+    "@sveltejs/package": "^2.3.10",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "premove": "^4.0.0",
-    "svelte": "^5.16.1"
+    "svelte": "^5.20.2"
   },
   "peerDependencies": {
     "svelte": "^5.0.0"
