@@ -1,6 +1,10 @@
 export * from '@tanstack/form-core'
 
+export { useStore } from '@tanstack/svelte-store'
+
 export { createForm, type SvelteFormApi } from './createForm.svelte.js'
 
-export type { Field } from './Field.svelte'
-export { createField } from './createField.svelte.js'
+// @ts-ignore tsc doesn't know about named exports from svelte files
+export { default as Field, createField } from './Field.svelte'
+
+export type { CreateField, FieldComponent } from './types.js'
